+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## 0.1.15 - 2023-09-19
 
-- Add support for a `trackEvaluation` callback for forwarding evaluation events to analytics tools,
-  etc.
+- Add support for a `afterEvaluationCallback` callback for forwarding evaluation events to analytics
+  tools, etc.
 
 ## 0.1.14 - 2023-07-11